refactor(contexts): Move context files into `contexts` directory. Remove unused files
